+++ logs/pact.log
@@ -1,0 +1,188 @@
+I, [2019-11-07T14:55:44.407883 #2320]  INFO -- : Registered expected interaction GET /heroes/42
+D, [2019-11-07T14:55:44.408910 #2320] DEBUG -- : {
+  "description": "a GET request for a hero",
+  "providerState": "a hero exists",
+  "request": {
+    "method": "GET",
+    "path": "/heroes/42",
+    "headers": {
+      "Accept": {
+        "json_class": "Pact::Term",
+        "data": {
+          "generate": "application/json; charset=utf-8",
+          "matcher": {"json_class":"Regexp","o":0,"s":"application\\/json; *charset=utf-8"}
+        }
+      }
+    }
+  },
+  "response": {
+    "status": 200,
+    "headers": {
+      "Content-Type": {
+        "json_class": "Pact::Term",
+        "data": {
+          "generate": "application/json; charset=utf-8",
+          "matcher": {"json_class":"Regexp","o":0,"s":"application\\/json; *charset=utf-8"}
+        }
+      }
+    },
+    "body": {
+      "json_class": "Pact::SomethingLike",
+      "contents": {
+        "id": 42,
+        "name": "Superman",
+        "superpower": "flying",
+        "universe": "DC"
+      }
+    }
+  }
+}
+I, [2019-11-07T14:55:44.440918 #2320]  INFO -- : Received request GET /heroes/42
+D, [2019-11-07T14:55:44.440918 #2320] DEBUG -- : {
+  "path": "/heroes/42",
+  "query": "",
+  "method": "get",
+  "headers": {
+    "Accept": "application/json; charset=utf-8",
+    "Referer": "http://localhost/",
+    "User-Agent": "Mozilla/5.0 (win32) AppleWebKit/537.36 (KHTML, like Gecko) jsdom/11.12.0",
+    "Accept-Language": "en",
+    "Origin": "http://localhost",
+    "Host": "localhost:8082",
+    "Accept-Encoding": "gzip, deflate",
+    "Connection": "close",
+    "Version": "HTTP/1.1"
+  }
+}
+I, [2019-11-07T14:55:44.441882 #2320]  INFO -- : Found matching response for GET /heroes/42
+D, [2019-11-07T14:55:44.441882 #2320] DEBUG -- : {
+  "status": 200,
+  "headers": {
+    "Content-Type": {
+      "json_class": "Pact::Term",
+      "data": {
+        "generate": "application/json; charset=utf-8",
+        "matcher": {
+          "json_class": "Regexp",
+          "o": 0,
+          "s": "application\\/json; *charset=utf-8"
+        }
+      }
+    }
+  },
+  "body": {
+    "json_class": "Pact::SomethingLike",
+    "contents": {
+      "id": 42,
+      "name": "Superman",
+      "superpower": "flying",
+      "universe": "DC"
+    }
+  }
+}
+I, [2019-11-07T14:55:44.505884 #2320]  INFO -- : Verifying - interactions matched
+I, [2019-11-07T14:55:44.516888 #2320]  INFO -- : Cleared interactions
+I, [2019-11-07T14:55:44.528884 #2320]  INFO -- : Registered expected interaction POST /heroes
+D, [2019-11-07T14:55:44.528884 #2320] DEBUG -- : {
+  "description": "a POST request to create a hero",
+  "providerState": "provider allows hero creation",
+  "request": {
+    "method": "POST",
+    "path": "/heroes",
+    "headers": {
+      "Accept": {
+        "json_class": "Pact::Term",
+        "data": {
+          "generate": "application/json; charset=utf-8",
+          "matcher": {"json_class":"Regexp","o":0,"s":"application\\/json; *charset=utf-8"}
+        }
+      },
+      "Content-Type": {
+        "json_class": "Pact::Term",
+        "data": {
+          "generate": "application/json; charset=utf-8",
+          "matcher": {"json_class":"Regexp","o":0,"s":"application\\/json; *charset=utf-8"}
+        }
+      }
+    },
+    "body": {
+      "name": "Superman",
+      "superpower": "flying",
+      "universe": "DC"
+    }
+  },
+  "response": {
+    "status": 201,
+    "headers": {
+      "Content-Type": {
+        "json_class": "Pact::Term",
+        "data": {
+          "generate": "application/json; charset=utf-8",
+          "matcher": {"json_class":"Regexp","o":0,"s":"application\\/json; *charset=utf-8"}
+        }
+      }
+    },
+    "body": {
+      "json_class": "Pact::SomethingLike",
+      "contents": {
+        "id": 42,
+        "name": "Superman",
+        "superpower": "flying",
+        "universe": "DC"
+      }
+    }
+  }
+}
+I, [2019-11-07T14:55:44.550912 #2320]  INFO -- : Received request POST /heroes
+D, [2019-11-07T14:55:44.551886 #2320] DEBUG -- : {
+  "path": "/heroes",
+  "query": "",
+  "method": "post",
+  "body": {
+    "name": "Superman",
+    "superpower": "flying",
+    "universe": "DC"
+  },
+  "headers": {
+    "Content-Length": "57",
+    "Content-Type": "application/json;charset=utf-8",
+    "Accept": "application/json; charset=utf-8",
+    "Referer": "http://localhost/",
+    "User-Agent": "Mozilla/5.0 (win32) AppleWebKit/537.36 (KHTML, like Gecko) jsdom/11.12.0",
+    "Accept-Language": "en",
+    "Origin": "http://localhost",
+    "Host": "localhost:8082",
+    "Accept-Encoding": "gzip, deflate",
+    "Connection": "close",
+    "Version": "HTTP/1.1"
+  }
+}
+I, [2019-11-07T14:55:44.552886 #2320]  INFO -- : Found matching response for POST /heroes
+D, [2019-11-07T14:55:44.552886 #2320] DEBUG -- : {
+  "status": 201,
+  "headers": {
+    "Content-Type": {
+      "json_class": "Pact::Term",
+      "data": {
+        "generate": "application/json; charset=utf-8",
+        "matcher": {
+          "json_class": "Regexp",
+          "o": 0,
+          "s": "application\\/json; *charset=utf-8"
+        }
+      }
+    }
+  },
+  "body": {
+    "json_class": "Pact::SomethingLike",
+    "contents": {
+      "id": 42,
+      "name": "Superman",
+      "superpower": "flying",
+      "universe": "DC"
+    }
+  }
+}
+I, [2019-11-07T14:55:44.571884 #2320]  INFO -- : Verifying - interactions matched
+I, [2019-11-07T14:55:44.583913 #2320]  INFO -- : Cleared interactions
+I, [2019-11-07T14:55:44.594889 #2320]  INFO -- : Updating pact for hero-provider at C:/Users/garciamo/PoCs/pact/pacts/hero-consumer-hero-provider.json
