+++ pacts/hero-consumer-hero-provider.json
@@ -1,0 +1,101 @@
+{
+  "consumer": {
+    "name": "hero-consumer"
+  },
+  "provider": {
+    "name": "hero-provider"
+  },
+  "interactions": [
+    {
+      "description": "a GET request for a hero",
+      "providerState": "a hero exists",
+      "request": {
+        "method": "GET",
+        "path": "/heroes/42",
+        "headers": {
+          "Accept": "application/json; charset=utf-8"
+        },
+        "matchingRules": {
+          "$.headers.Accept": {
+            "match": "regex",
+            "regex": "application\\/json; *charset=utf-8"
+          }
+        }
+      },
+      "response": {
+        "status": 200,
+        "headers": {
+          "Content-Type": "application/json; charset=utf-8"
+        },
+        "body": {
+          "id": 42,
+          "name": "Superman",
+          "superpower": "flying",
+          "universe": "DC"
+        },
+        "matchingRules": {
+          "$.headers.Content-Type": {
+            "match": "regex",
+            "regex": "application\\/json; *charset=utf-8"
+          },
+          "$.body": {
+            "match": "type"
+          }
+        }
+      }
+    },
+    {
+      "description": "a POST request to create a hero",
+      "providerState": "provider allows hero creation",
+      "request": {
+        "method": "POST",
+        "path": "/heroes",
+        "headers": {
+          "Accept": "application/json; charset=utf-8",
+          "Content-Type": "application/json; charset=utf-8"
+        },
+        "body": {
+          "name": "Superman",
+          "superpower": "flying",
+          "universe": "DC"
+        },
+        "matchingRules": {
+          "$.headers.Accept": {
+            "match": "regex",
+            "regex": "application\\/json; *charset=utf-8"
+          },
+          "$.headers.Content-Type": {
+            "match": "regex",
+            "regex": "application\\/json; *charset=utf-8"
+          }
+        }
+      },
+      "response": {
+        "status": 201,
+        "headers": {
+          "Content-Type": "application/json; charset=utf-8"
+        },
+        "body": {
+          "id": 42,
+          "name": "Superman",
+          "superpower": "flying",
+          "universe": "DC"
+        },
+        "matchingRules": {
+          "$.headers.Content-Type": {
+            "match": "regex",
+            "regex": "application\\/json; *charset=utf-8"
+          },
+          "$.body": {
+            "match": "type"
+          }
+        }
+      }
+    }
+  ],
+  "metadata": {
+    "pactSpecification": {
+      "version": "2.0.0"
+    }
+  }
+}
